Move class outside database folder
